pyproject.toml
[project]
name = "ansible-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Valentine Shatravenko", email = "vshatravenko@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pydot>=4.0.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
ansible-analyzer = "ansible_analyzer.main:main"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
